move doser to module folder
